auth0 integrated in App.js
