+++ client/src/components/App.js
@@ -29,4 +29,58 @@ function App() {
   );
 }
 
+class App extends Component {
+	goTo(route) {
+		this.props.history.replace(`/${route}`);
+	}
+
+	login() {
+		this.props.auth.login();
+	}
+
+	logout() {
+		this.props.auth.logout();
+	}
+
+	componentDidMount() {
+		const { renewSession } = this.props.auth;
+
+		if (localStorage.getItem('isLoggedIn') === 'true') {
+			renewSession();
+		}
+	}
+
+	render() {
+		const { isAuthenticated } = this.props.auth;
+
+		return (  
+			<div>
+				<Navbar fluid>
+					<Navbar.Header>
+						<Navbar.Brand>
+							<a href="#">Auth0 - React</a>
+						</Navbar.Brand>
+						<Button bsStyle="primary" className="btn-margin" onClick={this.goTo.bind(this, 'home')}>
+							Home
+						</Button>
+						{!isAuthenticated() &&
+							<Button bsStyle="primary" className="btn-margin" onClick={this.login.bind(this)}>
+								Log In
+							</Button>}
+						{isAuthenticated() &&
+							<Button bsStyle="primary" className="btn-margin" onClick={this.logout.bind(this)}>
+								Log Out
+							</Button>}
+						{isAuthenticated() &&
+							userHasScopes(['write:messages']) &&
+							<Button bsStyle="primary" className="btn-margin" onClick={this.goTo.bind(this, 'admin')}>
+								Admin
+							</Button>};
+					</Navbar.Header>
+				</Navbar>
+			</div>
+		);
+	}
+}
+
 export default App;
