+++ client/src/containers/Auth-Zero/Auth/Auth.js
@@ -1,0 +1,107 @@
+import history from '../../../history';
+import auth0 from 'auth0-js';
+import { AUTH_CONFIG } from './auth0-variables';
+
+export default class Auth {
+	accessToken;
+	idToken;
+	expiresAt;
+
+	// auth configuration
+	auth0 = new auth0.WebAuth({
+		domain: AUTH_CONFIG.domain,
+		clientID: AUTH_CONFIG.clientId,
+		redirectUri: AUTH_CONFIG.callbackUrl,
+		responseType: 'token id_token',
+		scope: 'openid'
+	});
+
+	constructor() {
+		this.login = this.login.bind(this);
+		this.logout = this.logout.bind(this);
+		this.handleAuthentication = this.handleAuthentication.bind(this);
+		this.isAuthenticated = this.isAuthenticated.bind(this);
+		this.getAccessToken = this.getAccessToken.bind(this);
+		this.getIdToken = this.getIdToken.bind(this);
+		this.renewSession = this.renewSession.bind(this);
+	}
+
+	// handle login built in function from Auth0
+	login() {
+		this.auth0.authorize();
+	}
+
+	// invoke setSession once user is authenticated
+	handleAuthentication() {
+		this.auth0.parseHash((err, authResult) => {
+			if (authResult && authResult.accessToken && authResult.idToken) {
+				this.setSession(authResult);
+			} else if (err) {
+				history.replace('/home');
+				console.log(err);
+				alert(`Error: ${err.error}. Check the console for further details.`);
+			}
+		});
+	}
+
+	getAccessToken() {
+		return this.accessToken;
+	}
+
+	getIdToken() {
+		return this.idToken;
+	}
+
+	// set token exp time
+	setSession(authResult) {
+		// Set isLoggedIn flag in localStorage
+		localStorage.setItem('isLoggedIn', 'true');
+
+		// Set the time that the access token will expire at
+		let expiresAt = authResult.expiresIn * 1000 + new Date().getTime();
+		this.accessToken = authResult.accessToken;
+		this.idToken = authResult.idToken;
+		this.expiresAt = expiresAt;
+
+		// navigate to the home route
+		history.replace('/home');
+	}
+
+	// schedule token renewal
+	renewSession() {
+		this.auth0.checkSession({}, (err, authResult) => {
+			if (authResult && authResult.accessToken && authResult.idToken) {
+				this.setSession(authResult);
+			} else if (err) {
+				this.logout();
+				console.log(err);
+				alert(`Could not get a new token (${err.error}: ${err.error_description}).`);
+			}
+		});
+	}
+
+	// logout user
+	logout() {
+		// Remove tokens and expiry time
+		this.accessToken = null;
+		this.idToken = null;
+		this.expiresAt = 0;
+
+		// Remove isLoggedIn flag from localStorage
+		localStorage.removeItem('isLoggedIn');
+
+		this.auth0.logout({
+			returnTo: window.location.origin
+		});
+
+		// navigate to the home route
+		history.replace('/home');
+	}
+
+	isAuthenticated() {
+		// Check whether the current time is past the
+		// access token's expiry time
+		let expiresAt = this.expiresAt;
+		return new Date().getTime() < expiresAt;
+	}
+}
