+++ client/src/containers/Auth-Zero/Auth/auth0-variables.js
@@ -1,0 +1,5 @@
+export const AUTH_CONFIG = {
+  domain: process.env.REACT_APP_AUTH0_DOMAIN,
+  clientId: process.env.REACT_APP_AUTH0_CLIENT_ID,
+  callbackUrl: process.env.REACT_APP_AUTH0_CALLBACK_URL
+};
