+++ client/src/containers/Auth-Zero/History.js
@@ -1,0 +1,3 @@
+const createHistory = require("history").createBrowserHistory
+
+export default createHistory()
