--- client/src/views/App/Auth/Auth.js
@@ -1,146 +1,0 @@
-import history from './History';
-import auth0 from 'auth0-js';
-import { AUTH_CONFIG } from './auth0-variables';
-
-export default class Auth {
-	accessToken;
-	idToken;
-	expiresAt;
-	userProfile;
-	scopes;
-	requestedScopes = 'openid profile read:messages write:messages';
-	tokenRenewalTimeout;
-
-	auth0 = new auth0.WebAuth({
-		domain: AUTH_CONFIG.domain,
-		clientID: AUTH_CONFIG.clientId,
-		redirectUri: AUTH_CONFIG.callbackUrl,
-		responseType: 'token id_token',
-		scope: 'openid',
-		scope: 'openid profile',
-		audience: 'http://shouldilivehere.auth0.com/api/v2',  // test link
-		scope: 'openid profile read:messages',
-		scope: this.requestedScopes
-	});
-
-	constructor() {
-		this.login = this.login.bind(this);
-		this.logout = this.logout.bind(this);
-		this.handleAuthentication = this.handleAuthentication.bind(this);
-		this.isAuthenticated = this.isAuthenticated.bind(this);
-		this.getAccessToken = this.getAccessToken.bind(this);
-		this.getIdToken = this.getIdToken.bind(this);
-		this.renewSession = this.renewSession.bind(this);
-		this.getProfile = this.getProfile.bind(this);
-		this.scheduleRenewal(); // token renewal
-	}
-
-	login() {
-		this.auth0.authorize();
-	}
-
-	handleAuthentication() {
-		this.auth0.parseHash((errir, authResult) => {
-			if (authResult && authResult.accessToken && authResult.idToken) {
-				this.setSession(authResult);
-				this.scheduleRenewal(); // schedule token
-			} else if (err) {
-				history.replace('/home');
-				console.log(err);
-				alert(`Error: ${err.error}. Check the console for further details.`);
-			}
-		});
-	}
-
-	getAccessToken() {
-		return this.accessToken;
-	}
-
-	getIdToken() {
-		return this.idToken;
-	}
-
-	setSession(authResult) {
-		// Set isLoggedIn flag in localStorage
-		localStorage.setItem('isLoggedIn', 'true');
-
-		// Set the time that the access token will expire at
-		let expiresAt = authResult.expiresIn * 1000 + new Date().getTime();
-		this.accessToken = authResult.accessToken;
-		this.idToken = authResult.idToken;
-		this.expiresAt = expiresAt;
-		this.scopes = authResult.scope || this.requestedScopes || '';
-
-		// navigate to the home route
-		history.replace('/home');
-	}
-
-	// token renewal
-	renewSession() {  
-		this.auth0.checkSession({}, (err, authResult) => {
-			if (authResult && authResult.accessToken && authResult.idToken) {
-				this.setSession(authResult);
-			} else if (err) {
-				this.logout();
-				console.log(err);
-				alert(`Could not get a new token (${err.error}: ${err.error_description}).`);
-			}
-		});
-	}
-
-	getProfile(cb) {
-		this.auth0.client.userInfo(this.accessToken, (err, profile) => {
-			if (profile) {
-				this.userProfile = profile;
-			}
-			cb(err, profile);
-		});
-	}
-
-	logout() {
-		// Remove tokens and expiry time
-		this.accessToken = null;
-		this.idToken = null;
-		this.expiresAt = 0;
-		this.profile = null;
-
-		// Remove isLoggedIn flag from localStorage
-		localStorage.removeItem('isLoggedIn');
-
-		this.auth0.logout({
-			returnTo: window.location.origin
-		});
-
-		// navigate to the home route
-		history.replace('/home');
-
-		clearTimeout(this.tokenRenewalTimeout); // cancel token renewal after logout
-	}
-
-	isAuthenticated() {
-		// Check whether the current time is past the
-		// access token's expiry time
-		let expiresAt = this.expiresAt;
-		return new Date().getTime() < expiresAt;
-	}
-
-	userHasScopes(scopes) {
-		const grantedScopes = this.scopes.split(' ');
-		return scopes.every(scope => grantedScopes.includes(scope));
-	}
-
-	// token renewal
-	scheduleRenewal() {
-		let expiresAt = this.expiresAt;
-		const timeout = expiresAt - Date.now();
-		if (timeout > 0) {
-		  this.tokenRenewalTimeout = setTimeout(() => {
-			this.renewSession();
-		  }, timeout);
-		}
-	  }
-	
-	  getExpiryDate() {
-		return JSON.stringify(new Date(this.expiresAt));
-	  }
-	}
