--- client/src/views/App/Auth/History.js
@@ -1,3 +1,0 @@
-import createHistory from 'history/createBrowserHistory'
-
-export default createHistory()
