--- client/src/views/App/Auth/auth0-variables.js
@@ -1,5 +1,0 @@
-export const AUTH_CONFIG = {  // needs to be customized - currently using test things
-    domain: '{DOMAIN}',
-    clientId: '{CLIENT_ID}',
-    callbackUrl: 'http://localhost:3000/callback'
-  }
